Edited all enemy scripts and prefabs, changed GameLevel1 to work with enemies, changed Project Settings
I edited the enemy scripts so they can actually damage the player now. I put all walls and floor on the Obstacles layer and made a NavMesh for the GameLevel1 scene. I changed the project settings so enemies don't collide with the player, nor each other, though that can be changed back later if desired.